+++ src/Console/Command/Command.php
@@ -15,7 +15,7 @@ declare(strict_types = 1);
 namespace Origin\Console\Command;
 
 use Origin\Core\HookTrait;
-use Origin\Model\ModelTrait;
+use Origin\Core\ModelTrait;
 
 use Origin\Console\ConsoleIo;
 use Origin\Console\CommandRunner;